--- SistemaGestaoHortifruti/src/Controller/ProdutosController.java
@@ -1,37 +1,0 @@
-package Controller;
-
-import Model.Produtos;
-import Repository.ProdutosRepository;
-
-public class ProdutosController implements ProdutosRepository {
-
-    @Override
-    public void pesquisaIdProduto(int id) {
-
-    }
-
-    @Override
-    public void pesquisaNomeProduto(String nome) {
-
-    }
-
-    @Override
-    public void listarProdutos() {
-
-    }
-
-    @Override
-    public void cadastrarProdutos(Produtos produto) {
-
-    }
-
-    @Override
-    public void atualizaPreco(Produtos preco) {
-
-    }
-
-    @Override
-    public void deletarProduto(String nome) {
-
-    }
-}
